+++ utils/zerovec/src/ule/mod.rs
@@ -14,14 +14,14 @@ mod chars;
 #[cfg(doc)]
 pub mod custom;
 mod encode;
-mod pair;
+mod tuple;
 mod plain;
 mod slices;
 
 pub use super::ZeroVecError;
 pub use chars::CharULE;
 pub use encode::{encode_varule_to_box, EncodeAsVarULE};
-pub use pair::PairULE;
+pub use tuple::PairULE;
 pub use plain::RawBytesULE;
 
 use alloc::alloc::Layout;
